# Introduce Journey::Ast to avoid extra ast walks