--- a/actionpack/test/journey/path/pattern_test.rb
+++ b/actionpack/test/journey/path/pattern_test.rb
@@ -133,22 +133,6 @@ module ActionDispatch
           assert_no_match(path, "/page/loving")
         end
 
-        def test_ast_sets_regular_expressions
-          requirements = { name: /(tender|love)/, value: /./ }
-          path = build_path(
-            "/page/:name/:value",
-            requirements,
-            SEPARATORS,
-            true
-          )
-
-          nodes = path.ast.grep(Nodes::Symbol)
-          assert_equal 2, nodes.length
-          nodes.each do |node|
-            assert_equal requirements[node.to_sym], node.regexp
-          end
-        end
-
         def test_match_data_with_group
           path = build_path(
             "/page/:name",