--- a/actionpack/test/journey/route_test.rb
+++ b/actionpack/test/journey/route_test.rb
@@ -24,7 +24,7 @@ module ActionDispatch
         path  = path_from_string "/:controller(/:action(/:id(.:format)))"
         route = Route.new(name: "name", app: app, path: path)
 
-        route.ast.grep(Nodes::Terminal).each do |node|
+        route.ast.root.grep(Nodes::Terminal).each do |node|
           assert_equal route, node.memo
         end
       end