@eugeneius 2020-07-27
My initial review:

 - This is really cool. 13% faster! 🏇 
 - How did you produce the benchmark results in the description? I'd like to see if I can replicate them on our app at work, which has a similar number of routes.
 - Are these new AST objects needed at runtime, or only while the routes are being defined? If it's the latter, I wonder how much more memory is now retained due to the memoized data. Maybe there's an opportunity to loop through the routes at the end and nil out their ASTs.

I'll take a closer look at the implementation soon.