@composerinteralia 2020-07-27
@eugeneius 

> How did you produce the benchmark results in the description? I'd like to see if I can replicate them on our app at work, which has a similar number of routes.

I added this to my routes file:

```diff
Rails.application.routes.draw do
+  before_object_count = GC.stat[:total_allocated_objects]
+  time = Benchmark.realtime do
    #... all the route definitions
+  end
+  puts GC.stat[:total_allocated_objects] - before_object_count
+  puts time
end
```

Then ran `bin/runner true` to see the output. The total allocated objects count was pretty consistent, so I didn't feel the need to do multiple runs there.

For the times I ran this not terribly elegant script, passing the refs I wanted to test out as arguments. You might want to change 500 to something more reasonable - I ran this and then went to bed.

```rb
def report(filename, ljust)
  times = File.read("journey_benchmark/#{filename}").lines.map(&:chomp).map(&:to_f)
  median = times.inject(&:+) / times.length
  dev = Math.sqrt((times.map { |time| (time - median)**2 }).inject(&:+) / (times.length - 1))

  puts "#{filename.ljust(ljust)} - #{median.round(3)} ± #{dev.round(3)} (#{times.length} runs)"
end

def run_benchmark(ref)
  puts "Running benchmark for '#{ref}'"

  Dir.chdir("path/to/local/rails") do
    `git checkout #{ref}`
  end

  500.times do 
    `bin/runner true | tail -1 >> "journey_benchmark/#{ref}"`
  end
end

Dir.chdir("path/to/application") do
  ARGV.each do |ref|
    run_benchmark(ref)
  end

  ljust = ARGV.map(&:length).max
  ARGV.each do |ref|
    report(ref, ljust)
  end
end
```

I'm sure there is a better way to do this...