@composerinteralia 2020-07-27
I put this in `RouteSet#eager_load`:

```rb
      def eager_load!
        router.eager_load!
        routes.each(&:eager_load!)
        require 'objspace'
        GC.start
        puts ObjectSpace.each_object(Object).count
        nil
      end
```

```rb
master         - 1141852
this commit - 1159326
```

Is this a reasonable way to measure what was retained?