@composerinteralia 2020-07-28
> Are these new AST objects needed at runtime, or only while the routes are being defined? If it's the latter, I wonder how much more memory is now retained due to the memoized data. Maybe there's an opportunity to loop through the routes at the end and nil out their ASTs.

Once possibility is to clear out the AST objects at the end of `RouteSet#eager_load!`. This would ensure that we don't retain these extra objects in any environment that eager loads (although we would still retain them in other environments, like the standard dev and test environments). But one concern we have here is whether there is ever a case where somebody would add routes **after** eager loading. If so, that could be a problem because adding a new route would [clear out the ast and simulator](https://github.com/rails/rails/blob/c6f02cbb56b6291533a3756bb5016964c76f5c20/actionpack/lib/action_dispatch/journey/routes.rb#L65-L77), at which point we would not have what we need to rebuild the simulator.