@casperisfine 2020-07-28
> Is this a reasonable way to measure what was retained?

Somewhat yes. But I'd suggest using http://github.com/SamSaffron/memory_profiler for a breakdown of what was retained etc.