@composerinteralia 2021-07-27
Rebased and reworked a little to take #41024 into consideration. I also wrote a cleaner benchmark to show that drawing routes is a bit faster with this patch:

Before:
```
Warming up --------------------------------------
drawing a simple route
                       435.000  i/100ms
drawing a complex route
                       306.000  i/100ms
Calculating -------------------------------------
drawing a simple route
                          4.067k (± 9.7%) i/s -     20.445k in   5.074692s
drawing a complex route
                          2.852k (± 6.2%) i/s -     14.382k in   5.063012s
```

After:
```
Warming up --------------------------------------
drawing a simple route
                       517.000  i/100ms
drawing a complex route
                       393.000  i/100ms
Calculating -------------------------------------
drawing a simple route
                          4.725k (± 6.2%) i/s -     23.782k in   5.052018s
drawing a complex route
                          3.532k (± 5.0%) i/s -     17.685k in   5.019272s
```

<details>
<summary>Benchmark</summary>

```rb
# frozen_string_literal: true

require "bundler/inline"

gemfile(true) do
  source "https://rubygems.org"

  git_source(:github) { |repo| "https://github.com/#{repo}.git" }

  gem "rails", path: "/Users/composerinteralia/code/rails/rails"
  gem "benchmark-ips"
  gem "stackprof"
  gem "memory_profiler"
end

require "action_controller"

class FooController < ActionController::Base; end

result = Benchmark.ips do |x|
  x.report("drawing a simple route") {
    rs_simple = ActionDispatch::Routing::RouteSet.new
    rs_simple.draw do
      get "/path", to: "controller#action"
    end
    rs_simple.eager_load!
  }

  x.report("drawing a complex route") {
    rs_simple = ActionDispatch::Routing::RouteSet.new
    rs_simple.draw do
      get "/path/(optional)/:symbol/with-:constraint/*glob", constraints: { constraint: /[abc]+/}, to: "controller#action"
    end
    rs_simple.eager_load!
  }
end
```
</details>